refactor(auth): 토큰 변경 감지 로직 공통 함수(onTokenChange)로 분리 및 전역 초기화 구조 정리 (#41)
- onTokenChange 유틸 추가 (AUTH_TOKEN_CHANGED / storage 이벤트 공통 감지)
- GlobalAuthHydrator: 토큰 기반 사용자 정보 하이드레이트 로직 개선
- AppInitializer: Authorization 헤더 동기화 로직 정리
- setupAuthTokenSync 및 tokenStore 간 역할 명확화
