+++ src/utils/auth/setupAuthTokenSync.ts
@@ -1,7 +1,8 @@
 // src/utils/auth/setupAuthTokenSync.ts
 "use client";
-import { tokenStore, AUTH_TOKEN_CHANGED, ACCESS_TOKEN_KEY } from "@/utils/auth/token.store";
+import { tokenStore } from "@/utils/auth/token.store";
 import { apiClient } from "@/apis/_client";
+import { onTokenChange } from "./onTokenChange";
 
 /**
  * setupAuthTokenSync
@@ -12,24 +13,9 @@ import { apiClient } from "@/apis/_client";
 
 export function setupAuthTokenSync() {
   if (typeof window === "undefined") return () => {};
-
   // 토큰 동기화 함수
   const sync = () => apiClient.setAuthToken(tokenStore.get() ?? undefined);
-
   // 초기 1회 동기화
   sync();
-
-  // 이벤트 리스너 등록
-  const onChanged = sync as EventListener;
-  const onStorage = (e: StorageEvent) => {
-    if (e.key === ACCESS_TOKEN_KEY) sync();
-  };
-  window.addEventListener(AUTH_TOKEN_CHANGED, onChanged);
-  window.addEventListener("storage", onStorage);
-
-  // 정리 함수 반환 (StrictMode 중복 마운트/언마운트 대비)
-  return () => {
-    window.removeEventListener(AUTH_TOKEN_CHANGED, onChanged);
-    window.removeEventListener("storage", onStorage);
-  };
+  return onTokenChange(sync); // 토큰 변할 때마다 헤더 갱신
 }
